requirements.txt
dedalus-labs
python-dotenv
flask
flask-cors